requirements.txt
# Core Django
Django==4.2.7
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# REST API Framework
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Image handling
Pillow==10.1.0

# API Documentation
drf-spectacular==0.26.5

# File uploads & cloud storage
django-storages==1.14.2
boto3==1.34.0  # For AWS S3 (optional)

# Email functionality
django-anymail==10.2

# Utilities
python-decouple==3.8
django-extensions==3.2.3

# Development tools
django-debug-toolbar==4.2.0

# Pagination
djangorestframework-recursive==0.1.2

# Validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.24

# Cache (optional)
redis==5.0.1
django-redis==5.4.0

# Task queue (optional)
celery==5.3.4

# CSV export
django-import-export==3.3.1

# Testing
factory-boy==3.3.0